correlation-analyzer/requirements.txt
pandas>=2.0.0
numpy>=2.0.0
scipy>=1.10.0
psycopg2-binary>=2.9.0
transformers>=4.30.0
torch>=2.4.0